Moved logic of methods
