+++ Website/flaskr/gestione_utente/GestioneUtenteController.py
@@ -1,18 +1,13 @@
 from flask import Blueprint, request, flash, session
-from flask_login import login_user, logout_user, login_required, current_user
-from werkzeug.security import check_password_hash, generate_password_hash
+from flask_login import login_user, logout_user, login_required
+from werkzeug.security import check_password_hash
 
 from Website.flaskr.Routes import home, area_personale, modifica_dati_utente_page, login_page, \
     registrazione_page
 from Website.flaskr.gestione_utente.GestioneUtenteService import get_apicoltore_by_email, get_cliente_by_email, \
-    controlla_email_esistente, registra_utente, modifica_profilo_personale
-from Website.flaskr.model.Apicoltore import Apicoltore
-from Website.flaskr.model.Cliente import Cliente
+    registra_utente, modifica_profilo_personale
 
 gu = Blueprint('gu', __name__)
-email_valida = r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b'
-spec = ["$", "#", "@", "!", "*", "£", "%", "&", "/", "(", ")", "=", "|",
-        "+", "-", "^", "_", "-", "?", ",", ":", ";", ".", "§", "°", "[", "]"]
 
 
 @gu.route('/login', methods=['GET', 'POST'])
@@ -61,25 +56,10 @@ def registrazione():
         email = request.form.get('email')
         pwd = request.form.get('password')
         cpwd = request.form.get('cpwd')
-        is_apicoltore = bool(int(request.form.get('isApicoltore')))
+        is_apicoltore = request.form.get('isApicoltore')
 
-        if controlla_campi(nome, cognome, indirizzo, citta, cap, telefono, email):
-            if not controlla_email_esistente(email):
-                flash("Email già esistente", category="error")
-            elif controlla_password(pwd, cpwd):
-                if is_apicoltore:
-                    user = Apicoltore(nome=nome, cognome=cognome, indirizzo=indirizzo, citta=citta, cap=cap,
-                                      telefono=telefono,
-                                      email=email, assistenza=0, password=generate_password_hash(pwd, method='sha256'))
-                else:
-                    user = Cliente(nome=nome, cognome=cognome, indirizzo=indirizzo, citta=citta, cap=cap,
-                                   telefono=telefono,
-                                   email=email, password=generate_password_hash(pwd, method='sha256'))
-
-                registra_utente(user)
-                session['isApicoltore'] = is_apicoltore
-                login_user(user)
-                return home()
+        if registra_utente(nome, cognome, indirizzo, citta, cap, telefono, email, pwd, cpwd, is_apicoltore):
+            return home()
     return registrazione_page()
 
 
@@ -97,61 +77,7 @@ def modifica_dati_utente():
         pwd = request.form.get('nuova_psw')
         cpwd = request.form.get('nuova_ripeti_psw')
 
-        if controlla_campi(nome, cognome, indirizzo, citta, cap, telefono, email):
-            if not controlla_email_esistente(email) and email != current_user.email:
-                flash("Email già esistente", category="error")
-            elif pwd != '':
-                if controlla_password(pwd, cpwd):
-                    modifica_profilo_personale(nome, cognome, email, telefono, citta, cap, indirizzo, pwd)
-                    return area_personale()
-            else:
-                modifica_profilo_personale(nome, cognome, email, telefono, citta, cap, indirizzo, pwd)
-                return area_personale()
+        if modifica_profilo_personale(nome, cognome, email, telefono, citta, cap, indirizzo, pwd, cpwd):
+            return area_personale()
+
     return modifica_dati_utente_page()
-
-
-def controlla_campi(nome, cognome, indirizzo, citta, cap, telefono, email):
-    if not 0 < len(nome) <= 45:
-        flash("Nome non valido", category="error")
-    elif not 0 < len(cognome) <= 45:
-        flash("Cognome non valido", category="error")
-    elif not 0 < len(indirizzo) <= 50:
-        flash("Indirizzo non valido", category="error")
-    elif not 0 < len(citta) <= 45:
-        flash("Città non valida", category="error")
-    elif not 0 < len(cap) <= 5 or not cap.isdigit():
-        flash("CAP non valido", category="error")
-    elif not 0 < len(telefono) <= 10 or not telefono.isdigit():
-        flash("Numero telefono non valido", category="error")
-    elif not 0 < len(email) <= 45:
-        flash("Email non valida", category="error")
-    else:
-        return True
-    return False
-
-
-def controlla_password(pwd, cpwd):
-    if len(pwd) < 8:
-        flash("Lunghezza password deve essere almeno 8 caratteri.", category="error")
-    elif not (controllo_caratteri_speciali(pwd) and controllo_numeri(pwd)):
-        flash("Inserire nel campo password almeno un carattere speciale ed un numero.", category="error")
-    elif pwd != cpwd:
-        flash("Password e Conferma Password non combaciano.", category="error")
-    else:
-        return True
-    return False
-
-
-def controllo_caratteri_speciali(psw):
-    for char in psw:
-        for symbol in spec:
-            if char == symbol:
-                return True
-    return False
-
-
-def controllo_numeri(psw):
-    for char in psw:
-        if char.isdigit():
-            return True
-    return False
