+++ Website/flaskr/templates/base.html
@@ -52,9 +52,6 @@
                     <li><a class="dropdown-item" href="{{url_for('views.inserimento_alveare_page')}}">Inserisci
                         Alveare</a></li>
                     <li><a class="dropdown-item"
-                           href="{{url_for('views.mostra_alveari', apicoltore_id=current_user.id)}}">Mostra i tuoi Alveari</a></li>
-
-                    <li><a class="dropdown-item"
                            href="{{url_for('views.mostra_articoli_in_vendita', apicoltore_id=current_user.id)}}">Mostra i tuoi Prodotti</a></li>
                     {% else %}
                     <li><a class="dropdown-item"
