+++ Website/flaskr/templates/inserimento_alveare.html
@@ -121,6 +121,7 @@
                                                 min="0"
                                                 name="prezzo"
                                                 type="number"
+                                                step="0.01"
                                                 required=true
                                                 placeholder="Inserisci prezzo"
                                         />
