+++ Website/flaskr/templates/inserimento_prodotto.html
@@ -229,6 +229,7 @@
                                                     id="prezzo"
                                                     name="prezzo"
                                                     type="number"
+                                                    step="0.01"
                                                     required=true
                                                     placeholder="Inserisci prezzo"
                                             />
